requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-huggingface
huggingface_hub